Rename Sessions route from /v2/fleet to /v2/sessions
The sidebar tab is labelled "Sessions" but the URL was /v2/fleet. Move
the route to /v2/sessions to match, and redirect the legacy /v2/fleet
paths (index and $sessionId) so existing bookmarks and external links
keep working.
