+++ src/routes/v2/sessions.index.tsx
@@ -1,0 +1,7 @@
+import { createFileRoute } from "@tanstack/react-router"
+
+import { FleetPage } from "@/components/V2/Fleet/FleetPage"
+
+export const Route = createFileRoute("/v2/sessions/")({
+  component: FleetPage,
+})
