+++ src/routes/v2/sessions.tsx
@@ -1,0 +1,16 @@
+import { createFileRoute, Outlet } from "@tanstack/react-router"
+
+export const Route = createFileRoute("/v2/sessions")({
+  component: SessionsLayout,
+  head: () => ({
+    meta: [
+      {
+        title: "Taskforce | Sessions",
+      },
+    ],
+  }),
+})
+
+function SessionsLayout() {
+  return <Outlet />
+}
